+++ constcheck/_main.py
@@ -31,13 +31,13 @@ def main(
     :key ignore_files: List of str objects for paths to exclude.
     :key ignore_from: Dict with list of str objects for strings to
         exclude from a particular path.
-    :key no_color: Boolean value to disable color output.
+    :key no_ansi: Boolean value to disable color output.
     :return: Exit status.
     """
     (
         path,
         values,
-        no_color,
+        no_ansi,
         string,
         ignore_strings,
         ignore_files,
@@ -46,7 +46,7 @@ def main(
     return _constcheck(
         path=path,
         values=values,
-        no_color=no_color,
+        no_ansi=no_ansi,
         string=string,
         ignore_strings=ignore_strings,
         ignore_files=ignore_files,